renamed TaskManagerService to TaskManagerGRPCService
